vpnext-news-backend/requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.2
groq>=0.9.0 
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
python-dotenv==1.0.1
schedule==1.2.2
httpx==0.27.0
